moving utils out from workers into separate package
